+++ src/main/java/com/project/planb/common/security/dto/TokenResDto.java
@@ -1,4 +1,4 @@
-package com.project.planb.domain.member.dto.res;
+package com.project.planb.common.security.dto;
 
 public record TokenResDto(
         // 성공 시 클라이언트 반환 dto
