+++ src/main/java/com/project/planb/domain/member/service/MemberService.java
@@ -1,8 +1,8 @@
 package com.project.planb.domain.member.service;
 
-import com.project.planb.domain.member.dto.req.MemberJoinReqDto;
-import com.project.planb.domain.member.dto.req.MemberLoginReqDto;
-import com.project.planb.domain.member.dto.res.TokenResDto;
+import com.project.planb.domain.member.dto.MemberJoinReqDto;
+import com.project.planb.domain.member.dto.MemberLoginReqDto;
+import com.project.planb.common.security.dto.TokenResDto;
 import com.project.planb.domain.member.entity.Member;
 import com.project.planb.common.exception.CustomException;
 import com.project.planb.common.exception.ErrorCode;
